+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/MavenAppEngineFacetUtil.java
@@ -6,6 +6,8 @@ import java.util.List;
 import java.util.Map;
 import org.apache.maven.model.Dependency;
 import org.eclipse.core.runtime.IProgressMonitor;
+
+import com.google.cloud.tools.eclipse.appengine.facets.standard.AppEngineStandardFacet;
 import com.google.cloud.tools.eclipse.util.MavenUtils;
 
 public class MavenAppEngineFacetUtil {
